refactor: TodoList Model 생성 및  ViewModel과 분리
Model 역할을 할 TodoList 구조체 생성
ViewModel에서 직접 모델에 접근하는 부분 분리
